src/ingestion_service/requirements.txt
requests
pika
sseclient-py
